requirements.txt
tensorflow==2.10.0
Django==4.1.2
numpy==1.23.3
Pillow==9.2.0
requests==2.28.1
gunicorn==20.1.0
django-heroku==0.3.1
dj-database-url==1.3.0